Use a blocking task for the final action too
